+++ src/gui/operation.rs
@@ -1,3 +1,4 @@
+use super::dia::open_file_chooser;
 use super::guii::Size2Di;
 use super::image_area::{open_uri_for_image_file, spawn_render_image};
 use super::zoom::{next_zoom_in_level, next_zoom_out_level};
@@ -120,15 +121,5 @@ pub(crate) fn ope_open_uri_for_image_file(uri_str: &str) {
 
 //
 pub(crate) fn ope_open_file_chooser_dialog() {
-    UI_GLOBAL.with(|global| {
-        if let Some((ref my_data, _)) = *global.borrow() {
-            {
-                let a_my_data = my_data.borrow();
-                if let Some(filename) = a_my_data.fc.run_open_file() {
-                    let uri = format!("file:///{}", filename);
-                    ope_open_uri_for_image_file(uri.as_str());
-                }
-            }
-        }
-    });
+    open_file_chooser();
 }