+++ CHANGELOG.md
@@ -2,6 +2,12 @@ aki-image-view-gtk TBD
 ===
 Unreleased changes. Release notes have not yet been written.
 
+0.2.3 (2021-08-21)
+=====
+
+* i separate ImVm.glade to Dialog.glade
+* fix: a render thread loop bug
+
 0.2.2 (2021-08-19)
 =====